Add files via upload
Yesterday, I added what was close to the final version of the HTML I handed in for my "Web Fundamentals" exam for Coding Dojo. Today, I am attaching the exact version of what I handed in, along with the stylesheet. I will definitely be editing this further in the future when I acquire more skill, but it is definitely a starting point... and one that I hope passed!